+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.3.44](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.3.43...@kong-ui-public/analytics-metric-provider@11.3.44) (2026-06-22)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [11.3.43](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.3.42...@kong-ui-public/analytics-metric-provider@11.3.43) (2026-06-19)